requirements.txt
torch>=2.0
botorch>=0.10
gpytorch>=1.11
pyyaml>=6.0
numpy>=1.24
scipy>=1.10
pandas>=2.0
matplotlib>=3.7
jinja2>=3.1
langgraph>=0.2
openai>=1.0
paramiko>=3.4
